requirements.txt
python-telegram-bot>=22.0
python-dotenv>=1.0.0
openai>=1.0.0
chromadb>=0.4.0
datasets>=2.0.0